create doi minting app and move some logic there
