--- apps/services/__init__.py
@@ -1,3 +1,0 @@
-from .invenio_service import InvenioService
-
-__all__ = ["save_metadata_to_invenio_then_mint_doi", "InvenioService"]
