+++ doi_minting/urls.py
@@ -1,0 +1,7 @@
+app_name = "doi_minting"
+
+from django.urls import path
+
+urlpatterns = [
+    # path('mint/', views.mint_doi, name='mint_doi'),
+]
